# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DuckyAPI",
    "description": "A customer facing api for WildDuck",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [{ "url": "http://localhost:3000" }],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
...
      "DomainAlias": {
        "type": "object",
        "properties": {
          "domain": {
            "type": "string",
            "example": "example.com",
            "description": "The domain name"
          },
          "dkim": {
            "type": "boolean",
            "example": false,
            "readOnly": true,
            "description": "If DKIM is active for this domain"
          }
        },
        "required": ["domain"]
      },
      "Domain": {
        "type": "object",
        "properties": {
          "domain": {
            "type": "string",
            "example": "example.com",
            "description": "The domain name"
          },
          "dkim": {
            "type": "boolean",
            "example": false,
            "readOnly": true,
            "description": "If DKIM is active for this domain"
          },
          "admin": {
            "type": "boolean",
            "example": true,
            "readOnly": true,
            "description": "If this user is the domain admin, this currently serves no function"
          },
          "aliases": {
            "description": "Domains aliased to this domain",
            "readOnly": true,
            "type": "array",
            "items": { "$ref": "#/components/schemas/DomainAlias" }
          }
        },
        "required": ["domain"]
      },
      "DnsCheckMxRecord": {
        "type": "object",
        "properties": {
          "exchange": {
            "type": "string",
            "example": "mx.example.com",
            "description": "MX record server"
          },
          "priority": {
            "type": "number",
            "example": 1,
            "description": "MX record priority"
          }
        },
        "required": ["exchange", "priority"]
      },
      "DnsCheckDkimRecord": {
        "type": "object",
        "properties": {
          "selector": {
            "type": "string",
            "example": "default",
            "description": "DKIM record selector"
          },
          "value": {
            "type": "string",
            "example": "v=DKIM1;t=s;p=MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAseRvI//jDgRsZ1BtGixcLO16/B8yEzsgVSBvCWgwf39LRAey14eZLoyyolX7wVUe71VN67cEuey7XlYGHzGntDtLh/CmI8vvDaiym0VNv8zrZok2TbYW0I4Ts9YkNtCUC5EKjyrwX7AT97ZjiXVX6JK+oEmdtgwxtrQc9+trYj3udlStEmpH0yluY3kSmUYDe3e4TEdLUX7+x/i4D8+65dIXdw52cRNka9aMpH7ZdsfPvrFd6y+ItOuX1Zsb8uFdQz21/Tf1aVczwbZgpUFfpyt55erLwfFLdlH7aRwBIJGQDMzl4SFkGgxDuSPjUePHO266PiHm2/r8A0515n3ZCwIDAQAB",
            "description": "DKIM record value"
          }
        },
        "required": ["selector", "value"]
      },
      "DnsCheckCurrentValues": {
        "type": "object",
        "properties": {
          "mx": {
            "description": "List of DNS records",
            "type": "array",
            "items": { "$ref": "#/components/schemas/DnsCheckMxRecord" }
          },
          "spf": {
            "type": "string",
            "example": "v=spf1 include:example.com -all",
            "description": "Value of the SPF record"
          },
          "dkim": {
            "allOf": [
              { "$ref": "#/components/schemas/DnsCheckDkimRecord" },
              { "description": "DKIM record selector and value" }
            ]
          }
        },
        "required": ["mx", "spf"]
      },
      "DnsCheckCorrectValues": {
        "type": "object",
        "properties": {
          "mx": {
            "description": "List of DNS records",
            "type": "array",
            "items": { "$ref": "#/components/schemas/DnsCheckMxRecord" }
          },
          "spf": {
            "type": "string",
            "example": "v=spf1 include:example.com -all",
            "description": "Value of the SPF record"
          },
          "dkim": {
            "allOf": [
              { "$ref": "#/components/schemas/DnsCheckDkimRecord" },
              { "description": "DKIM record selector and value" }
            ]
          }
        },
        "required": ["mx", "spf"]
      },
      "DnsCheckError": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "dkim",
            "description": "Type of error/warning. Can be ns, mx, spf, dkim"
          },
          "error": {
            "type": "string",
            "example": "DkimNotFound",
            "description": "Machine readable error/warning string"
          },
          "message": {
            "type": "string",
            "example": "DKIM is enabled, but no record was found",
            "description": "Human readable error/warning message"
          }
        },
        "required": ["type", "error", "message"]
      },
      "DnsCheck": {
        "type": "object",
        "properties": {
          "currentValues": {
            "allOf": [
              { "$ref": "#/components/schemas/DnsCheckCurrentValues" },
              { "description": "Current values of the DNS records" }
            ]
          },
          "correctValues": {
            "allOf": [
              { "$ref": "#/components/schemas/DnsCheckCorrectValues" },
              { "description": "Correct values of the DNS records" }
            ]
          },
          "errors": {
            "description": "List of errors with the DNS records",
            "type": "array",
            "items": { "$ref": "#/components/schemas/DnsCheckError" }
          },
          "warnings": {
            "description": "List of warnings with the DNS records",
            "type": "array",
            "items": { "$ref": "#/components/schemas/DnsCheckError" }
          }
        },
        "required": ["currentValues", "correctValues", "errors", "warnings"]
      },
...
    }
  },
  "paths": {
...
    "/domains/{domain}": {
      "delete": {
        "operationId": "deleteDomain",
        "summary": "Delete a domain",
        "description": "WARNING: This will also delete any email accounts, forwarders, and DKIM keys associated with this domain",
        "parameters": [
          {
            "name": "domain",
            "required": true,
            "in": "path",
            "description": "example.com",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Domain successfully deleted" },
          "400": { "description": "Bad user input" },
          "401": { "description": "Invalid or expired token" },
          "404": { "description": "Domain not found on account" }
        },
        "tags": ["Domains"],
        "security": [{ "bearer": [] }],
        "x-code-samples": [
          {
            "lang": "Shell",
            "label": "Curl",
            "source": "curl --request DELETE \\\n  --url http://localhost:3000/domains/%7Bdomain%7D \\\n  --header 'authorization: Bearer REPLACE_BEARER_TOKEN'"
          },
          {
            "lang": "JavaScript",
            "label": "Node.js",
            "source": "var http = require(\"http\");\n\nvar options = {\n  \"method\": \"DELETE\",\n  \"hostname\": \"localhost\",\n  \"port\": \"3000\",\n  \"path\": \"/domains/%7Bdomain%7D\",\n  \"headers\": {\n    \"authorization\": \"Bearer REPLACE_BEARER_TOKEN\"\n  }\n};\n\nvar req = http.request(options, function (res) {\n  var chunks = [];\n\n  res.on(\"data\", function (chunk) {\n    chunks.push(chunk);\n  });\n\n  res.on(\"end\", function () {\n    var body = Buffer.concat(chunks);\n    console.log(body.toString());\n  });\n});\n\nreq.end();"
          },
          {
            "lang": "JavaScript",
            "label": "JavaScript",
            "source": "var data = null;\n\nvar xhr = new XMLHttpRequest();\nxhr.withCredentials = true;\n\nxhr.addEventListener(\"readystatechange\", function () {\n  if (this.readyState === this.DONE) {\n    console.log(this.responseText);\n  }\n});\n\nxhr.open(\"DELETE\", \"http://localhost:3000/domains/%7Bdomain%7D\");\nxhr.setRequestHeader(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\");\n\nxhr.send(data);"
          },
          {
            "lang": "Python",
            "label": "Python3",
            "source": "import http.client\n\nconn = http.client.HTTPConnection(\"localhost:3000\")\n\nheaders = { 'authorization': \"Bearer REPLACE_BEARER_TOKEN\" }\n\nconn.request(\"DELETE\", \"/domains/%7Bdomain%7D\", headers=headers)\n\nres = conn.getresponse()\ndata = res.read()\n\nprint(data.decode(\"utf-8\"))"
          },
          {
            "lang": "PHP",
            "label": "PHP",
            "source": "<?php\n\n$curl = curl_init();\n\ncurl_setopt_array($curl, array(\n  CURLOPT_PORT => \"3000\",\n  CURLOPT_URL => \"http://localhost:3000/domains/%7Bdomain%7D\",\n  CURLOPT_RETURNTRANSFER => true,\n  CURLOPT_ENCODING => \"\",\n  CURLOPT_MAXREDIRS => 10,\n  CURLOPT_TIMEOUT => 30,\n  CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,\n  CURLOPT_CUSTOMREQUEST => \"DELETE\",\n  CURLOPT_HTTPHEADER => array(\n    \"authorization: Bearer REPLACE_BEARER_TOKEN\"\n  ),\n));\n\n$response = curl_exec($curl);\n$err = curl_error($curl);\n\ncurl_close($curl);\n\nif ($err) {\n  echo \"cURL Error #:\" . $err;\n} else {\n  echo $response;\n}"
          },
          {
            "lang": "Java",
            "label": "Java",
            "source": "HttpResponse<String> response = Unirest.delete(\"http://localhost:3000/domains/%7Bdomain%7D\")\n  .header(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\")\n  .asString();"
          },
          {
            "lang": "C#",
            "label": "C#",
            "source": "var client = new RestClient(\"http://localhost:3000/domains/%7Bdomain%7D\");\nvar request = new RestRequest(Method.DELETE);\nrequest.AddHeader(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\");\nIRestResponse response = client.Execute(request);"
          },
          {
            "lang": "C",
            "label": "C",
            "source": "CURL *hnd = curl_easy_init();\n\ncurl_easy_setopt(hnd, CURLOPT_CUSTOMREQUEST, \"DELETE\");\ncurl_easy_setopt(hnd, CURLOPT_URL, \"http://localhost:3000/domains/%7Bdomain%7D\");\n\nstruct curl_slist *headers = NULL;\nheaders = curl_slist_append(headers, \"authorization: Bearer REPLACE_BEARER_TOKEN\");\ncurl_easy_setopt(hnd, CURLOPT_HTTPHEADER, headers);\n\nCURLcode ret = curl_easy_perform(hnd);"
          }
        ]
      }
    },
    "/domains": {
      "get": {
        "operationId": "getDomains",
        "summary": "List domains",
        "parameters": [],
        "responses": {
          "200": {
            "description": "A list of domains",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Domain" }
                }
              }
            }
          },
          "400": { "description": "Bad user input" },
          "401": { "description": "Invalid or expired token" }
        },
        "tags": ["Domains"],
        "security": [{ "bearer": [] }],
        "x-code-samples": [
          {
            "lang": "Shell",
            "label": "Curl",
            "source": "curl --request GET \\\n  --url http://localhost:3000/domains \\\n  --header 'authorization: Bearer REPLACE_BEARER_TOKEN'"
          },
          {
            "lang": "JavaScript",
            "label": "Node.js",
            "source": "var http = require(\"http\");\n\nvar options = {\n  \"method\": \"GET\",\n  \"hostname\": \"localhost\",\n  \"port\": \"3000\",\n  \"path\": \"/domains\",\n  \"headers\": {\n    \"authorization\": \"Bearer REPLACE_BEARER_TOKEN\"\n  }\n};\n\nvar req = http.request(options, function (res) {\n  var chunks = [];\n\n  res.on(\"data\", function (chunk) {\n    chunks.push(chunk);\n  });\n\n  res.on(\"end\", function () {\n    var body = Buffer.concat(chunks);\n    console.log(body.toString());\n  });\n});\n\nreq.end();"
          },
          {
            "lang": "JavaScript",
            "label": "JavaScript",
            "source": "var data = null;\n\nvar xhr = new XMLHttpRequest();\nxhr.withCredentials = true;\n\nxhr.addEventListener(\"readystatechange\", function () {\n  if (this.readyState === this.DONE) {\n    console.log(this.responseText);\n  }\n});\n\nxhr.open(\"GET\", \"http://localhost:3000/domains\");\nxhr.setRequestHeader(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\");\n\nxhr.send(data);"
          },
          {
            "lang": "Python",
            "label": "Python3",
            "source": "import http.client\n\nconn = http.client.HTTPConnection(\"localhost:3000\")\n\nheaders = { 'authorization': \"Bearer REPLACE_BEARER_TOKEN\" }\n\nconn.request(\"GET\", \"/domains\", headers=headers)\n\nres = conn.getresponse()\ndata = res.read()\n\nprint(data.decode(\"utf-8\"))"
          },
          {
            "lang": "PHP",
            "label": "PHP",
            "source": "<?php\n\n$curl = curl_init();\n\ncurl_setopt_array($curl, array(\n  CURLOPT_PORT => \"3000\",\n  CURLOPT_URL => \"http://localhost:3000/domains\",\n  CURLOPT_RETURNTRANSFER => true,\n  CURLOPT_ENCODING => \"\",\n  CURLOPT_MAXREDIRS => 10,\n  CURLOPT_TIMEOUT => 30,\n  CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,\n  CURLOPT_CUSTOMREQUEST => \"GET\",\n  CURLOPT_HTTPHEADER => array(\n    \"authorization: Bearer REPLACE_BEARER_TOKEN\"\n  ),\n));\n\n$response = curl_exec($curl);\n$err = curl_error($curl);\n\ncurl_close($curl);\n\nif ($err) {\n  echo \"cURL Error #:\" . $err;\n} else {\n  echo $response;\n}"
          },
          {
            "lang": "Java",
            "label": "Java",
            "source": "HttpResponse<String> response = Unirest.get(\"http://localhost:3000/domains\")\n  .header(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\")\n  .asString();"
          },
          {
            "lang": "C#",
            "label": "C#",
            "source": "var client = new RestClient(\"http://localhost:3000/domains\");\nvar request = new RestRequest(Method.GET);\nrequest.AddHeader(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\");\nIRestResponse response = client.Execute(request);"
          },
          {
            "lang": "C",
            "label": "C",
            "source": "CURL *hnd = curl_easy_init();\n\ncurl_easy_setopt(hnd, CURLOPT_CUSTOMREQUEST, \"GET\");\ncurl_easy_setopt(hnd, CURLOPT_URL, \"http://localhost:3000/domains\");\n\nstruct curl_slist *headers = NULL;\nheaders = curl_slist_append(headers, \"authorization: Bearer REPLACE_BEARER_TOKEN\");\ncurl_easy_setopt(hnd, CURLOPT_HTTPHEADER, headers);\n\nCURLcode ret = curl_easy_perform(hnd);"
          }
        ]
      },
      "post": {
        "operationId": "addDomain",
        "summary": "Add a domain",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Domain" }
            }
          }
        },
        "responses": {
          "201": { "description": "Domain successfully added" },
          "400": { "description": "Bad user input" },
          "401": { "description": "Invalid or expired token" }
        },
        "tags": ["Domains"],
        "security": [{ "bearer": [] }],
        "x-code-samples": [
          {
            "lang": "Shell",
            "label": "Curl",
            "source": "curl --request POST \\\n  --url http://localhost:3000/domains \\\n  --header 'authorization: Bearer REPLACE_BEARER_TOKEN' \\\n  --header 'content-type: application/json' \\\n  --data '{\"domain\":\"example.com\"}'"
          },
          {
            "lang": "JavaScript",
            "label": "Node.js",
            "source": "var http = require(\"http\");\n\nvar options = {\n  \"method\": \"POST\",\n  \"hostname\": \"localhost\",\n  \"port\": \"3000\",\n  \"path\": \"/domains\",\n  \"headers\": {\n    \"content-type\": \"application/json\",\n    \"authorization\": \"Bearer REPLACE_BEARER_TOKEN\"\n  }\n};\n\nvar req = http.request(options, function (res) {\n  var chunks = [];\n\n  res.on(\"data\", function (chunk) {\n    chunks.push(chunk);\n  });\n\n  res.on(\"end\", function () {\n    var body = Buffer.concat(chunks);\n    console.log(body.toString());\n  });\n});\n\nreq.write(JSON.stringify({domain: 'example.com'}));\nreq.end();"
          },
          {
            "lang": "JavaScript",
            "label": "JavaScript",
            "source": "var data = JSON.stringify({\n  \"domain\": \"example.com\"\n});\n\nvar xhr = new XMLHttpRequest();\nxhr.withCredentials = true;\n\nxhr.addEventListener(\"readystatechange\", function () {\n  if (this.readyState === this.DONE) {\n    console.log(this.responseText);\n  }\n});\n\nxhr.open(\"POST\", \"http://localhost:3000/domains\");\nxhr.setRequestHeader(\"content-type\", \"application/json\");\nxhr.setRequestHeader(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\");\n\nxhr.send(data);"
          },
          {
            "lang": "Python",
            "label": "Python3",
            "source": "import http.client\n\nconn = http.client.HTTPConnection(\"localhost:3000\")\n\npayload = \"{\\\"domain\\\":\\\"example.com\\\"}\"\n\nheaders = {\n    'content-type': \"application/json\",\n    'authorization': \"Bearer REPLACE_BEARER_TOKEN\"\n    }\n\nconn.request(\"POST\", \"/domains\", payload, headers)\n\nres = conn.getresponse()\ndata = res.read()\n\nprint(data.decode(\"utf-8\"))"
          },
          {
            "lang": "PHP",
            "label": "PHP",
            "source": "<?php\n\n$curl = curl_init();\n\ncurl_setopt_array($curl, array(\n  CURLOPT_PORT => \"3000\",\n  CURLOPT_URL => \"http://localhost:3000/domains\",\n  CURLOPT_RETURNTRANSFER => true,\n  CURLOPT_ENCODING => \"\",\n  CURLOPT_MAXREDIRS => 10,\n  CURLOPT_TIMEOUT => 30,\n  CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,\n  CURLOPT_CUSTOMREQUEST => \"POST\",\n  CURLOPT_POSTFIELDS => \"{\\\"domain\\\":\\\"example.com\\\"}\",\n  CURLOPT_HTTPHEADER => array(\n    \"authorization: Bearer REPLACE_BEARER_TOKEN\",\n    \"content-type: application/json\"\n  ),\n));\n\n$response = curl_exec($curl);\n$err = curl_error($curl);\n\ncurl_close($curl);\n\nif ($err) {\n  echo \"cURL Error #:\" . $err;\n} else {\n  echo $response;\n}"
          },
          {
            "lang": "Java",
            "label": "Java",
            "source": "HttpResponse<String> response = Unirest.post(\"http://localhost:3000/domains\")\n  .header(\"content-type\", \"application/json\")\n  .header(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\")\n  .body(\"{\\\"domain\\\":\\\"example.com\\\"}\")\n  .asString();"
          },
          {
            "lang": "C#",
            "label": "C#",
            "source": "var client = new RestClient(\"http://localhost:3000/domains\");\nvar request = new RestRequest(Method.POST);\nrequest.AddHeader(\"content-type\", \"application/json\");\nrequest.AddHeader(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\");\nrequest.AddParameter(\"application/json\", \"{\\\"domain\\\":\\\"example.com\\\"}\", ParameterType.RequestBody);\nIRestResponse response = client.Execute(request);"
          },
          {
            "lang": "C",
            "label": "C",
            "source": "CURL *hnd = curl_easy_init();\n\ncurl_easy_setopt(hnd, CURLOPT_CUSTOMREQUEST, \"POST\");\ncurl_easy_setopt(hnd, CURLOPT_URL, \"http://localhost:3000/domains\");\n\nstruct curl_slist *headers = NULL;\nheaders = curl_slist_append(headers, \"content-type: application/json\");\nheaders = curl_slist_append(headers, \"authorization: Bearer REPLACE_BEARER_TOKEN\");\ncurl_easy_setopt(hnd, CURLOPT_HTTPHEADER, headers);\n\ncurl_easy_setopt(hnd, CURLOPT_POSTFIELDS, \"{\\\"domain\\\":\\\"example.com\\\"}\");\n\nCURLcode ret = curl_easy_perform(hnd);"
          }
        ]
      }
    },
    "/domains/{domain}/DNS": {
      "get": {
        "operationId": "checkDNS",
        "summary": "Get and check DNS records",
        "parameters": [
          {
            "name": "domain",
            "required": true,
            "in": "path",
            "description": "example.com",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "The current and the correct DNS records for this domain",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DnsCheck" }
              }
            }
          },
          "400": { "description": "Bad user input" },
          "401": { "description": "Invalid or expired token" },
          "404": { "description": "Domain not found on account" }
        },
        "tags": ["Domains"],
        "security": [{ "bearer": [] }],
        "x-code-samples": [
          {
            "lang": "Shell",
            "label": "Curl",
            "source": "curl --request GET \\\n  --url http://localhost:3000/domains/%7Bdomain%7D/DNS \\\n  --header 'authorization: Bearer REPLACE_BEARER_TOKEN'"
          },
          {
            "lang": "JavaScript",
            "label": "Node.js",
            "source": "var http = require(\"http\");\n\nvar options = {\n  \"method\": \"GET\",\n  \"hostname\": \"localhost\",\n  \"port\": \"3000\",\n  \"path\": \"/domains/%7Bdomain%7D/DNS\",\n  \"headers\": {\n    \"authorization\": \"Bearer REPLACE_BEARER_TOKEN\"\n  }\n};\n\nvar req = http.request(options, function (res) {\n  var chunks = [];\n\n  res.on(\"data\", function (chunk) {\n    chunks.push(chunk);\n  });\n\n  res.on(\"end\", function () {\n    var body = Buffer.concat(chunks);\n    console.log(body.toString());\n  });\n});\n\nreq.end();"
          },
          {
            "lang": "JavaScript",
            "label": "JavaScript",
            "source": "var data = null;\n\nvar xhr = new XMLHttpRequest();\nxhr.withCredentials = true;\n\nxhr.addEventListener(\"readystatechange\", function () {\n  if (this.readyState === this.DONE) {\n    console.log(this.responseText);\n  }\n});\n\nxhr.open(\"GET\", \"http://localhost:3000/domains/%7Bdomain%7D/DNS\");\nxhr.setRequestHeader(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\");\n\nxhr.send(data);"
          },
          {
            "lang": "Python",
            "label": "Python3",
            "source": "import http.client\n\nconn = http.client.HTTPConnection(\"localhost:3000\")\n\nheaders = { 'authorization': \"Bearer REPLACE_BEARER_TOKEN\" }\n\nconn.request(\"GET\", \"/domains/%7Bdomain%7D/DNS\", headers=headers)\n\nres = conn.getresponse()\ndata = res.read()\n\nprint(data.decode(\"utf-8\"))"
          },
          {
            "lang": "PHP",
            "label": "PHP",
            "source": "<?php\n\n$curl = curl_init();\n\ncurl_setopt_array($curl, array(\n  CURLOPT_PORT => \"3000\",\n  CURLOPT_URL => \"http://localhost:3000/domains/%7Bdomain%7D/DNS\",\n  CURLOPT_RETURNTRANSFER => true,\n  CURLOPT_ENCODING => \"\",\n  CURLOPT_MAXREDIRS => 10,\n  CURLOPT_TIMEOUT => 30,\n  CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,\n  CURLOPT_CUSTOMREQUEST => \"GET\",\n  CURLOPT_HTTPHEADER => array(\n    \"authorization: Bearer REPLACE_BEARER_TOKEN\"\n  ),\n));\n\n$response = curl_exec($curl);\n$err = curl_error($curl);\n\ncurl_close($curl);\n\nif ($err) {\n  echo \"cURL Error #:\" . $err;\n} else {\n  echo $response;\n}"
          },
          {
            "lang": "Java",
            "label": "Java",
            "source": "HttpResponse<String> response = Unirest.get(\"http://localhost:3000/domains/%7Bdomain%7D/DNS\")\n  .header(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\")\n  .asString();"
          },
          {
            "lang": "C#",
            "label": "C#",
            "source": "var client = new RestClient(\"http://localhost:3000/domains/%7Bdomain%7D/DNS\");\nvar request = new RestRequest(Method.GET);\nrequest.AddHeader(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\");\nIRestResponse response = client.Execute(request);"
          },
          {
            "lang": "C",
            "label": "C",
            "source": "CURL *hnd = curl_easy_init();\n\ncurl_easy_setopt(hnd, CURLOPT_CUSTOMREQUEST, \"GET\");\ncurl_easy_setopt(hnd, CURLOPT_URL, \"http://localhost:3000/domains/%7Bdomain%7D/DNS\");\n\nstruct curl_slist *headers = NULL;\nheaders = curl_slist_append(headers, \"authorization: Bearer REPLACE_BEARER_TOKEN\");\ncurl_easy_setopt(hnd, CURLOPT_HTTPHEADER, headers);\n\nCURLcode ret = curl_easy_perform(hnd);"
          }
        ]
      }
    },
    "/domains/{domain}/aliases": {
      "post": {
        "operationId": "addAlias",
        "summary": "Add a domain alias",
        "parameters": [
          {
            "name": "domain",
            "required": true,
            "in": "path",
            "description": "example.com",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DomainAlias" }
            }
          }
        },
        "responses": {
          "201": { "description": "Alias successfully added" },
          "400": { "description": "Bad user input" },
          "401": { "description": "Invalid or expired token" },
          "404": { "description": "Domain not found on account" }
        },
        "tags": ["Domains"],
        "security": [{ "bearer": [] }],
        "x-code-samples": [
          {
            "lang": "Shell",
            "label": "Curl",
            "source": "curl --request POST \\\n  --url http://localhost:3000/domains/%7Bdomain%7D/aliases \\\n  --header 'authorization: Bearer REPLACE_BEARER_TOKEN' \\\n  --header 'content-type: application/json' \\\n  --data '{\"domain\":\"example.com\"}'"
          },
          {
            "lang": "JavaScript",
            "label": "Node.js",
            "source": "var http = require(\"http\");\n\nvar options = {\n  \"method\": \"POST\",\n  \"hostname\": \"localhost\",\n  \"port\": \"3000\",\n  \"path\": \"/domains/%7Bdomain%7D/aliases\",\n  \"headers\": {\n    \"content-type\": \"application/json\",\n    \"authorization\": \"Bearer REPLACE_BEARER_TOKEN\"\n  }\n};\n\nvar req = http.request(options, function (res) {\n  var chunks = [];\n\n  res.on(\"data\", function (chunk) {\n    chunks.push(chunk);\n  });\n\n  res.on(\"end\", function () {\n    var body = Buffer.concat(chunks);\n    console.log(body.toString());\n  });\n});\n\nreq.write(JSON.stringify({domain: 'example.com'}));\nreq.end();"
          },
          {
            "lang": "JavaScript",
            "label": "JavaScript",
            "source": "var data = JSON.stringify({\n  \"domain\": \"example.com\"\n});\n\nvar xhr = new XMLHttpRequest();\nxhr.withCredentials = true;\n\nxhr.addEventListener(\"readystatechange\", function () {\n  if (this.readyState === this.DONE) {\n    console.log(this.responseText);\n  }\n});\n\nxhr.open(\"POST\", \"http://localhost:3000/domains/%7Bdomain%7D/aliases\");\nxhr.setRequestHeader(\"content-type\", \"application/json\");\nxhr.setRequestHeader(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\");\n\nxhr.send(data);"
          },
          {
            "lang": "Python",
            "label": "Python3",
            "source": "import http.client\n\nconn = http.client.HTTPConnection(\"localhost:3000\")\n\npayload = \"{\\\"domain\\\":\\\"example.com\\\"}\"\n\nheaders = {\n    'content-type': \"application/json\",\n    'authorization': \"Bearer REPLACE_BEARER_TOKEN\"\n    }\n\nconn.request(\"POST\", \"/domains/%7Bdomain%7D/aliases\", payload, headers)\n\nres = conn.getresponse()\ndata = res.read()\n\nprint(data.decode(\"utf-8\"))"
          },
          {
            "lang": "PHP",
            "label": "PHP",
            "source": "<?php\n\n$curl = curl_init();\n\ncurl_setopt_array($curl, array(\n  CURLOPT_PORT => \"3000\",\n  CURLOPT_URL => \"http://localhost:3000/domains/%7Bdomain%7D/aliases\",\n  CURLOPT_RETURNTRANSFER => true,\n  CURLOPT_ENCODING => \"\",\n  CURLOPT_MAXREDIRS => 10,\n  CURLOPT_TIMEOUT => 30,\n  CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,\n  CURLOPT_CUSTOMREQUEST => \"POST\",\n  CURLOPT_POSTFIELDS => \"{\\\"domain\\\":\\\"example.com\\\"}\",\n  CURLOPT_HTTPHEADER => array(\n    \"authorization: Bearer REPLACE_BEARER_TOKEN\",\n    \"content-type: application/json\"\n  ),\n));\n\n$response = curl_exec($curl);\n$err = curl_error($curl);\n\ncurl_close($curl);\n\nif ($err) {\n  echo \"cURL Error #:\" . $err;\n} else {\n  echo $response;\n}"
          },
          {
            "lang": "Java",
            "label": "Java",
            "source": "HttpResponse<String> response = Unirest.post(\"http://localhost:3000/domains/%7Bdomain%7D/aliases\")\n  .header(\"content-type\", \"application/json\")\n  .header(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\")\n  .body(\"{\\\"domain\\\":\\\"example.com\\\"}\")\n  .asString();"
          },
          {
            "lang": "C#",
            "label": "C#",
            "source": "var client = new RestClient(\"http://localhost:3000/domains/%7Bdomain%7D/aliases\");\nvar request = new RestRequest(Method.POST);\nrequest.AddHeader(\"content-type\", \"application/json\");\nrequest.AddHeader(\"authorization\", \"Bearer REPLACE_BEARER_TOKEN\");\nrequest.AddParameter(\"application/json\", \"{\\\"domain\\\":\\\"example.com\\\"}\", ParameterType.RequestBody);\nIRestResponse response = client.Execute(request);"
          },
          {
            "lang": "C",
            "label": "C",
            "source": "CURL *hnd = curl_easy_init();\n\ncurl_easy_setopt(hnd, CURLOPT_CUSTOMREQUEST, \"POST\");\ncurl_easy_setopt(hnd, CURLOPT_URL, \"http://localhost:3000/domains/%7Bdomain%7D/aliases\");\n\nstruct curl_slist *headers = NULL;\nheaders = curl_slist_append(headers, \"content-type: application/json\");\nheaders = curl_slist_append(headers, \"authorization: Bearer REPLACE_BEARER_TOKEN\");\ncurl_easy_setopt(hnd, CURLOPT_HTTPHEADER, headers);\n\ncurl_easy_setopt(hnd, CURLOPT_POSTFIELDS, \"{\\\"domain\\\":\\\"example.com\\\"}\");\n\nCURLcode ret = curl_easy_perform(hnd);"
          }
        ]
      }
    },
...
  }
}